+++ src/client/Join.js
@@ -4,6 +4,7 @@ import { makeStyles } from "@material-ui/core/styles";
 import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
 import Typography from "@material-ui/core/Typography";
+import TopBar from "./components/topbar";
 import { addPlayer, checkValid, genID } from "./utils";
 
 const useStyles = makeStyles({
@@ -37,6 +38,7 @@ export default function Join({ location }) {
 
   return !done ? (
     <div>
+      <TopBar />
       <Typography variant="h3" className={classes.mar}>
         {" "}
         Welcome,
